+++ src/main/java/org/betonquest/betonquest/config/StandardPatchingConfigAccessor.java
@@ -1,6 +1,7 @@
 package org.betonquest.betonquest.config;
 
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
+import org.betonquest.betonquest.config.patcher.Patcher;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.plugin.Plugin;
 import org.jetbrains.annotations.Nullable;
